add import service

+++ auth/Model/Import.cs
@@ -5,14 +5,13 @@ namespace auth.Model
     public class Import
     {
         public int Id { get; set; }
-        public float Total { get; set; }
+        public int Total { get; set; }
         [ForeignKey("User")]
         public string UserId {get; set; }
         [ForeignKey("Supplier")]
         public int SupplierId { get; set; }
         public DateTime CreatedDate { get; set; } = DateTime.Now;
         public bool IsDeleted { get; set; } = false;
-
         public List<ImportDetail> Details { get; set; }
         #region
         public User User { get; set; }
